Updated membership request related API route naming to be RESTful.
